Move JVM specific tests to RTSSpec
